Create README.md
As part of my final project for the Data Journalism course, I analyzed statements from countries in the global north and global south and compared them to the final report at COP26. This GitHub repository is to provide readers with access to the steps I've taken during the analyses and see how I came to write my story.